requirements.txt
# --- Framework Web y Comunicación en Tiempo Real ---
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
gevent-websocket==0.10.1

# --- Lógica de Negocio (Tus dependencias originales) ---
bcrypt==4.1.3
colorama==0.4.6
openpyxl==3.1.2